1.5/project/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Cors==4.0.0
